Changed names of classes and created n-dimensional euler
