+++ hpc/Project_C++_Files/Functions2.h
@@ -4,3 +4,22 @@
 
 #pragma once
 
+#include <cmath>
+#include <cstdio>
+
+// Globals
+extern int A;
+extern int B;
+extern int k;
+
+double problemOde2(double* x, double u);
+
+double exactMinimumControl2(double* x);
+
+double costFunction2(double* x, double alpha);
+
+double driftFunction2(double* x, double alpha);
+
+double diffFunction2(double* x);
+
+double sigmaFunction2(double* x);
